rename emptyLineBeforeFixer to fixEmptyLineBefore
