+++ lib/rules/comment-empty-line-before/index.cjs
@@ -4,7 +4,7 @@
 
 const validateTypes = require('../../utils/validateTypes.cjs');
 const addEmptyLineBefore = require('../../utils/addEmptyLineBefore.cjs');
-const emptyLineBeforeFixer = require('../../utils/emptyLineBeforeFixer.cjs');
+const fixEmptyLineBefore = require('../../utils/fixEmptyLineBefore.cjs');
 const hasEmptyLine = require('../../utils/hasEmptyLine.cjs');
 const isAfterComment = require('../../utils/isAfterComment.cjs');
 const configurationComment = require('../../utils/configurationComment.cjs');
@@ -109,7 +109,7 @@ const rule = (primary, secondaryOptions, context) => {
 			const mutate = expectEmptyLineBefore ? addEmptyLineBefore : removeEmptyLinesBefore;
 
 			// Fix
-			const fix = emptyLineBeforeFixer.bind(null, {
+			const fix = fixEmptyLineBefore.bind(null, {
 				node: comment,
 				newline: context.newline,
 				mutate,
