+++ lib/rules/comment-empty-line-before/index.mjs
@@ -1,6 +1,6 @@
 import { isRegExp, isString } from '../../utils/validateTypes.mjs';
 import addEmptyLineBefore from '../../utils/addEmptyLineBefore.mjs';
-import fixer from '../../utils/emptyLineBeforeFixer.mjs';
+import fixer from '../../utils/fixEmptyLineBefore.mjs';
 import hasEmptyLine from '../../utils/hasEmptyLine.mjs';
 import isAfterComment from '../../utils/isAfterComment.mjs';
 import { isConfigurationComment } from '../../utils/configurationComment.mjs';
