+++ lib/rules/custom-property-empty-line-before/index.cjs
@@ -5,7 +5,7 @@
 const typeGuards = require('../../utils/typeGuards.cjs');
 const addEmptyLineBefore = require('../../utils/addEmptyLineBefore.cjs');
 const blockString = require('../../utils/blockString.cjs');
-const emptyLineBeforeFixer = require('../../utils/emptyLineBeforeFixer.cjs');
+const fixEmptyLineBefore = require('../../utils/fixEmptyLineBefore.cjs');
 const getPreviousNonSharedLineCommentNode = require('../../utils/getPreviousNonSharedLineCommentNode.cjs');
 const hasEmptyLine = require('../../utils/hasEmptyLine.cjs');
 const isAfterComment = require('../../utils/isAfterComment.cjs');
@@ -110,7 +110,7 @@ const rule = (primary, secondaryOptions, context) => {
 			const mutate = expectEmptyLineBefore ? addEmptyLineBefore : removeEmptyLinesBefore;
 
 			// Fix
-			const fix = emptyLineBeforeFixer.bind(null, {
+			const fix = fixEmptyLineBefore.bind(null, {
 				node: decl,
 				newline: context.newline,
 				mutate,
