+++ lib/rules/declaration-empty-line-before/index.mjs
@@ -1,7 +1,7 @@
 import { isAtRule, isRoot, isRule } from '../../utils/typeGuards.mjs';
 import addEmptyLineBefore from '../../utils/addEmptyLineBefore.mjs';
 import blockString from '../../utils/blockString.mjs';
-import fixer from '../../utils/emptyLineBeforeFixer.mjs';
+import fixer from '../../utils/fixEmptyLineBefore.mjs';
 import hasEmptyLine from '../../utils/hasEmptyLine.mjs';
 import isAfterComment from '../../utils/isAfterComment.mjs';
 import isAfterStandardPropertyDeclaration from '../../utils/isAfterStandardPropertyDeclaration.mjs';
