+++ lib/rules/rule-empty-line-before/index.cjs
@@ -3,7 +3,7 @@
 'use strict';
 
 const addEmptyLineBefore = require('../../utils/addEmptyLineBefore.cjs');
-const emptyLineBeforeFixer = require('../../utils/emptyLineBeforeFixer.cjs');
+const fixEmptyLineBefore = require('../../utils/fixEmptyLineBefore.cjs');
 const getPreviousNonSharedLineCommentNode = require('../../utils/getPreviousNonSharedLineCommentNode.cjs');
 const hasEmptyLine = require('../../utils/hasEmptyLine.cjs');
 const isAfterSingleLineComment = require('../../utils/isAfterSingleLineComment.cjs');
@@ -124,7 +124,7 @@ const rule = (primary, secondaryOptions, context) => {
 			const mutate = expectEmptyLineBefore ? addEmptyLineBefore : removeEmptyLinesBefore;
 
 			// Fix
-			const fix = emptyLineBeforeFixer.bind(null, {
+			const fix = fixEmptyLineBefore.bind(null, {
 				node: ruleNode,
 				newline: context.newline,
 				mutate,
