+++ lib/rules/rule-empty-line-before/index.mjs
@@ -1,5 +1,5 @@
 import addEmptyLineBefore from '../../utils/addEmptyLineBefore.mjs';
-import fixer from '../../utils/emptyLineBeforeFixer.mjs';
+import fixer from '../../utils/fixEmptyLineBefore.mjs';
 import getPreviousNonSharedLineCommentNode from '../../utils/getPreviousNonSharedLineCommentNode.mjs';
 import hasEmptyLine from '../../utils/hasEmptyLine.mjs';
 import isAfterSingleLineComment from '../../utils/isAfterSingleLineComment.mjs';
