+++ lib/utils/fixEmptyLineBefore.cjs
@@ -1,4 +1,8 @@
-import { assert } from './validateTypes.mjs';
+// NOTICE: This file is generated by Rollup. To modify it,
+// please instead edit the ESM counterpart and rebuild with Rollup (npm run build).
+'use strict';
+
+const validateTypes = require('./validateTypes.cjs');
 
 /** @typedef {import('stylelint').Range} Range */
 /** @typedef {import('postcss').Node} Node */
@@ -12,7 +16,7 @@ import { assert } from './validateTypes.mjs';
  * @param {Callback} o.mutate
  * @returns {Range}
  */
-export default function fixer({ node, newline, mutate }) {
+const fixer = ({ node, newline, mutate }) => {
 	const range = node.rangeBy({ word: node.toString() });
 	let count;
 
@@ -22,17 +26,19 @@ export default function fixer({ node, newline, mutate }) {
 			break;
 		case 'removeEmptyLinesBefore':
 			count = (node.raws.before.match(/\n/g) || []).length - 1;
-			assert(count > 0);
+			validateTypes.assert(count > 0);
 			break;
 		default:
 			throw new Error('unknown mutate function');
 	}
 
-	assert(newline);
+	validateTypes.assert(newline);
 	mutate(node, newline);
 
 	range.start.line -= count;
 	range.end.line -= count;
 
 	return range;
-}
+};
+
+module.exports = fixer;
