+++ lib/utils/fixEmptyLineBefore.mjs
@@ -1,8 +1,4 @@
-// NOTICE: This file is generated by Rollup. To modify it,
-// please instead edit the ESM counterpart and rebuild with Rollup (npm run build).
-'use strict';
-
-const validateTypes = require('./validateTypes.cjs');
+import { assert } from './validateTypes.mjs';
 
 /** @typedef {import('stylelint').Range} Range */
 /** @typedef {import('postcss').Node} Node */
@@ -16,7 +12,7 @@ const validateTypes = require('./validateTypes.cjs');
  * @param {Callback} o.mutate
  * @returns {Range}
  */
-function fixer({ node, newline, mutate }) {
+export default ({ node, newline, mutate }) => {
 	const range = node.rangeBy({ word: node.toString() });
 	let count;
 
@@ -26,19 +22,17 @@ function fixer({ node, newline, mutate }) {
 			break;
 		case 'removeEmptyLinesBefore':
 			count = (node.raws.before.match(/\n/g) || []).length - 1;
-			validateTypes.assert(count > 0);
+			assert(count > 0);
 			break;
 		default:
 			throw new Error('unknown mutate function');
 	}
 
-	validateTypes.assert(newline);
+	assert(newline);
 	mutate(node, newline);
 
 	range.start.line -= count;
 	range.end.line -= count;
 
 	return range;
-}
-
-module.exports = fixer;
+};
